feat: added image repo

+++ main.cpp
@@ -6,27 +6,73 @@
 #include "src/Thresholding.h"
 #include "src/MatTools.h"
 #include "src/History.h"
+#include "src/ImageRepo.h"
 
 using namespace std;
 using namespace cv;
 
-void showImage(Mat &image, const string& name) {
-    Mat &imageToPrint = image;
-//    resize(image, imageToPrint, Size(600, 600));
-//    resizeWindow(name, 600, 600);
-    imshow(name, imageToPrint);
-    waitKey(0);
-    destroyWindow(name);
+void printMask(const Mat &mask) {
+    cout << "Mask: " << endl;
+    cout << "\t[ " << endl;
+    for (int i = 0; i < mask.cols; ++i) {
+        for (int j = 0; j < mask.rows; ++j) {
+            int label = mask.at<uchar>(i, j);
+            cout << "\t" << label << " ";
+        }
+        cout << endl;
+    }
+
+    cout << "\t] " << endl;
+
 }
 
 
-void printImage(Mat &image, const string& name) {
+void showImage(Mat &image, const string &name, bool destroy = true) {
+    Mat &imageToPrint = image;
+    // resize without interpolation
+//    resize(image, imageToPrint, Size(600, 600), 0, 0, INTER_NEAREST);
+    imshow(name, imageToPrint);
+    waitKey(0);
+    if (destroy)
+        destroyWindow(name);
+}
+
+void printImageByMask(const Mat &mask, const Mat &originalImage) {
+    // create a new image L*a*b* with the same size of the mask
+    Mat imageLabThr = Mat::zeros(mask.size(), CV_32FC3);
+    auto itImageOriginal = originalImage.begin<Vec3f>();
+    auto itImage = imageLabThr.begin<Vec3f>();
+    auto itMask = mask.begin<uchar>();
+    auto itEnd = mask.end<uchar>();
+    while (itMask != itEnd) {
+        auto label = *itMask;
+        auto &pixel = *itImage;
+        auto imageStatsClassI = History::imageStatsMap[label];
+        pixel[0] = (*itImageOriginal)[0];
+//        pixel[0] = imageStatsClassI.mean.l;
+        pixel[1] = imageStatsClassI.mean.a;
+        pixel[2] = imageStatsClassI.mean.b;
+        itImageOriginal++;
+        itImage++;
+        itMask++;
+    }
+
+    Mat imageThr;
+    cvtColor(imageLabThr, imageThr, COLOR_Lab2BGR);
+    imageThr *= 255.0;
+    imageThr.convertTo(imageThr, CV_8UC3);
+
+    showImage(imageThr, "Thresholding");
+}
+
+
+void printImage(Mat &image, const string &name) {
     cout << "Image: " << name << endl;
     cout << "\tMatrix L:" << name << endl;
     cout << "\t[ " << endl;
     for (int i = 0; i < image.cols; ++i) {
         for (int j = 0; j < image.rows; ++j) {
-            cout << "\t"<< image.at<Vec3f>(i, j)[0];
+            cout << "\t" << image.at<Vec3f>(i, j)[0];
         }
         cout << endl;
     }
@@ -57,80 +103,39 @@ void printImage(Mat &image, const string& name) {
 }
 
 int main() {
-    // Image
-//    auto name = "img/c_3x3.png";
-//    auto name = "img/corazon.png";
-    auto name = "img/jaguarcito.png";
-    // Read image.
-    Mat image = imread(name);
-    // Convert to float values.
-    Mat imageFloat;
-    image.convertTo(imageFloat, CV_32FC3);
-    // Normalize.
-    imageFloat /= 255.0;
-    // Convert to BGR2Lab.
-    Mat imageLab;
-    cvtColor(imageFloat, imageLab, COLOR_BGR2Lab);
+    ImageRepo repo = ImageRepo();
+    Mat image;
+    Mat imageOriginal;
+//    repo.arcoiris_3x3(image, imageOriginal);
+    repo.jaguarcito(image, imageOriginal);
     // init mask with zeros.
-    Mat mask = Mat::zeros(imageLab.size(), CV_8UC1);
+    Mat mask = Mat::zeros(image.size(), CV_8UC1);
 
-    cout << "imageLab.size()" << imageLab.size() << endl;
+    cout << "image.size()" << image.size() << endl;
+    // print image
+//    printImage(imageLab, name);
+    string name = "Original";
+    showImage(imageOriginal, name, false);
 
-    vector<int> labels = {0};
-    vector<History> historyVector;
-    auto iHistory = History(mask, labels);
-    historyVector.push_back(iHistory);
+
     auto &stats = History::imageStatsMap;
+    vector<int> labels = {0};
+    MatTools::meanCovariance(image, mask, stats, labels);
+    printMask(mask);
+    printImageByMask(mask, image);
 
-    MatTools::meanCovariance(imageLab, mask, stats, labels);
-
-    printImage(imageLab, name);
-    showImage(image, name);
-
-    auto thresholding = Thresholding(imageLab);
-
-    for (int i = 0; i < 3; ++i) {
+    auto thresholding = Thresholding(image);
+    for (int i = 0; i < 20; ++i) {
         cout << "Level: " << i + 1 << endl;
 
-        if (labels.empty()) {
+        vector<int> nextLabels = thresholding.propagate(mask, labels);
+        if (nextLabels.empty()) {
             break;
         }
-         vector<int> nextLabels = thresholding.propagate(mask, labels);
-
-        // print labels
-        cout << "Labels: ";
-        for (const auto &itemLabels: nextLabels) {
-            cout << itemLabels << " ";
-        }
-        cout << endl;
-
-        iHistory = History(mask, nextLabels);
-        historyVector.push_back(iHistory);
+        printMask(mask);
+        printImageByMask(mask, image);
 
         labels = nextLabels;
-    }
-
-    for (const auto &itemHistoryVector: historyVector) {
-        Mat m = itemHistoryVector.mask;
-        Mat imageLabThr = imageLab.clone();
-        auto itImage = imageLabThr.begin<Vec3f>();
-        auto itMask = m.begin<uchar>();
-        auto itEnd = m.end<uchar>();
-        while (itMask != itEnd) {
-            auto imageStatsClassI = History::imageStatsMap[*itMask];
-            (*itImage)[0] = 29.0;
-            (*itImage)[1] = imageStatsClassI.mean.a;
-            (*itImage)[2] = imageStatsClassI.mean.b;
-            itImage++;
-            itMask++;
-        }
-
-        Mat imageThr;
-        cvtColor(imageLabThr, imageThr, COLOR_Lab2BGR);
-        imageThr *= 255.0;
-        imageThr.convertTo(imageThr, CV_8UC3);
-
-        showImage(imageThr, "Thresholding");
     }
 
     return 0;
